+++ CHANGELOG.md
@@ -1,15 +1,18 @@
-## 2.0.0 (Work in progress)
+## 2.0.0
 
 - Additions:
-  - Added the `meta` package for static analysis.
+  - Added developer dependencies:
+    - `meta` for static analysis.
+    - `test` for testing.
   - Added support for the 'Sitemap' field.
   - Added support for specifying:
     - The precedent rule type for determining whether a certain user-agent can
       or cannot access a certain path. (`PrecedentRuleType`)
     - The comparison strategy to use for comparing rule precedence.
       (`PrecedenceStrategy`)
+  - Added tests.
 - Changes:
-  - Bumped the minimum SDK version to `2.17.0`.
+  - Bumped the minimum SDK version to `2.17.0` for enhanced enum support.
 - Improvements:
   - Made all structs `const` and marked them as `@sealed` and `@immutable`.
 - Deletions: